+++ apps/port-labeler/src/features/i18n/i18nFlag.utils.ts
@@ -1,0 +1,6 @@
+import { useTranslation } from 'react-i18next'
+
+export const useI18nFlag = (iso: string) => {
+  const { t } = useTranslation('flags')
+  return t(iso as any)
+}
