+++ libs/ui-components/src/logo/Logo.types.ts
@@ -1,0 +1,7 @@
+export enum SubBrands {
+  MarineManager = 'Marine Manager',
+  CarrierVessels = 'Carrier Vessels',
+  PortLabeler = 'Port Labeler',
+}
+
+export type LogoTypes = 'default' | 'invert'
