+++ libs/ui-components/src/logo/index.ts
@@ -1,1 +1,2 @@
 export * from './Logo'
+export * from './Logo.types'
